pyproject.toml
[project]
name = "serial-mcp"
version = "0.1.0"
description = "Serial terminal emulator with MCP server"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26",
    "pyserial>=3.5",
    "uvicorn",
    "starlette",
    "anyio",
    "websockets",
    "pyyaml>=6.0",
]

[project.scripts]
serial-mcp = "serial_mcp.cli:main_sync"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
